requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
open3d>=0.17.0
numpy>=1.24.0
matplotlib>=3.7.0
h5py>=3.8.0
scikit-learn>=1.2.0
tqdm>=4.65.0
Pillow>=9.5.0
opencv-python>=4.7.0
